Move `HashingServiceImpl` to module runtime usable by all sandboxes - Reference `HashingService` instead of platform `DigestService` inside sandboxes
