+++ libs/ledger/ledger-common-data/src/main/kotlin/net/corda/ledger/common/data/transaction/WireTransaction.kt
@@ -2,7 +2,7 @@ package net.corda.ledger.common.data.transaction
 
 import net.corda.crypto.core.concatByteArrays
 import net.corda.crypto.core.toByteArray
-import net.corda.v5.cipher.suite.DigestService
+import net.corda.v5.application.crypto.HashingService
 import net.corda.v5.cipher.suite.merkle.MerkleTreeProvider
 import net.corda.v5.crypto.DigestAlgorithmName
 import net.corda.v5.crypto.SecureHash
@@ -18,7 +18,7 @@ import java.util.Objects
 
 class WireTransaction(
     private val merkleTreeProvider: MerkleTreeProvider,
-    private val digestService: DigestService,
+    private val digestService: HashingService,
     val privacySalt: PrivacySalt,
     val componentGroupLists: List<List<ByteArray>>,
     val metadata: TransactionMetadata
