+++ testing/ledger/ledger-common-testkit/src/main/kotlin/net/corda/ledger/common/testkit/WireTransactionExample.kt
@@ -4,8 +4,8 @@ import net.corda.common.json.validation.JsonValidator
 import net.corda.ledger.common.data.transaction.TransactionMetadata
 import net.corda.ledger.common.data.transaction.WireTransaction
 import net.corda.ledger.common.data.transaction.factory.WireTransactionFactory
+import net.corda.v5.application.crypto.HashingService
 import net.corda.v5.application.marshalling.JsonMarshallingService
-import net.corda.v5.cipher.suite.DigestService
 import net.corda.v5.cipher.suite.merkle.MerkleTreeProvider
 
 fun WireTransactionFactory.createExample(
@@ -24,7 +24,7 @@ fun WireTransactionFactory.createExample(
 
 @Suppress("LongParameterList")
 fun getWireTransactionExample(
-    digestService: DigestService,
+    digestService: HashingService,
     merkleTreeProvider: MerkleTreeProvider,
     jsonMarshallingService: JsonMarshallingService,
     jsonValidator: JsonValidator,
